connect to api
