+++ scrybe_frontend/src/pages/AgentReport/Content.jsx
@@ -4,57 +4,23 @@ import close from "./assets/icon.svg";
 import Charts from "./components/ChartContainer";
 import AgentDetails from "./components/AgentDetails";
 import { useState, useEffect } from "react";
-import axios from "axios";
-import { agentData } from "./components/Data";
 import { useAgentReport } from "./hooks";
+import { useAgentDetails } from "./hooks";
 
 const Content = () => {
   const agentReport = useAgentReport();
-
-  // chart data
-
-  const [agentDets, setAgentDets] = useState({});
-  const [data_id, setData_id] = useState("5");
-
-  useEffect(() => {
-    const data =
-      "grant_type=&username=rambeybello%40gmail.com&password=aaaaaaaa&scope=&client_id=&client_secret=";
-    axios.post("https://api.heed.hng.tech/login", data).then((res) => {
-      const headers = {
-        Authorization: `Bearer ${res.data.access_token}`,
-      };
-      axios
-        .get(
-          `https://api.heed.hng.tech/total-agent-analysis?agent_id=${data_id}`,
-          {
-            headers,
-          }
-        )
-        .then((res) => {
-          console.log(res.data);
-          setAgentDets(res.data);
-        });
-    });
-  }, []);
-
-  // console.log(agentDets);
-  // chart data
-
+  const agentDetail = useAgentDetails();
   const [selectReport, setSelectReport] = useState([]);
 
+  useEffect(() => {
+    setSelectReport(agentDetail.week);
+  }, [agentDetail]);
+
   const handleDate = (e) => {
-    setSelectReport(agentData[e.target.value]);
-    // setSelectReport(agentDets[e.target.value]);
+    setSelectReport(agentDetail[e.target.value]);
   };
 
-  useEffect(() => {
-    setSelectReport(agentData.week);
-    // setSelectReport(agentDets.week);
-    // console.log(selectReport);
-  }, []);
-
   return (
-    // <div className={styles.container}>
     <div className={styles.mainWrapper}>
       <div className={styles.header}>
         <h1>Agent Report</h1>
@@ -91,7 +57,6 @@ const Content = () => {
         <AgentDetails />
       </div>
     </div>
-    // </div>
   );
 };
