+++ scrybe_frontend/src/pages/AgentReport/styles/AgentReport.module.scss
@@ -115,6 +115,13 @@
 @media screen and (max-width: 888px) {
   .container {
     .mainWrapper {
+      width: 90%;
+
+      .header {
+        h1 {
+          font-size: 20px;
+        }
+      }
       .idcont {
         flex-direction: column;
         align-items: flex-start;
