--- scrybe_frontend/src/pages/EmailVerification/Email.module.scss
@@ -1,54 +1,0 @@
-.verify {
-    position: relative;
-    background: #fff;
-    height: 100vh;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    background: rgba(225, 225, 225, 0.66);
-    backdrop-filter: blur(4px) saturate(200%);
-
-    .popup {
-        max-width: 350px;
-        padding: 20px;
-        margin: auto;
-        text-align: center;
-        background: #fff;
-        border-radius: 10px;
-
-        img {
-            width: 80px;
-            margin: 0 auto 30px auto;
-        }
-
-        .status {
-            font-weight: 600;
-            font-size: 20px;
-            color: #006cff;
-            margin-bottom: 20px;
-        }
-
-        .msg {
-            margin-bottom: 20px;
-        }
-
-        a {
-            background: #006cff;
-            display: block;
-            width: 100%;
-            padding: 12px;
-            border-radius: 3px;
-            color: #fff;
-            transition: all .9s ease;
-            border: 1px solid transparent;
-
-            &:hover {
-                border-color: #006cff;
-                background: transparent;
-                color: #006cff;
-            }
-        }
-    }
-}
-
-
